wip: save current state [skip ci]
