+++ jaxley/utils/cell_utils.py
@@ -10,96 +10,6 @@ from jax.typing import ArrayLike
 
 from jaxley.utils.misc_utils import cumsum_leading_zero
 
-
-def surface_area_segments(ls, r1, r2, dr):
-    """Surface area of truncated cone segments."""
-    slant_height = np.sqrt(ls**2 + dr**2)
-    return 2 * np.pi * (r1 + r2) / 2 * slant_height
-
-
-def volume_segments(ls, r1, r2):
-    """Volume of truncated cone segments."""
-    return (np.pi * ls / 3) * (r1**2 + r1 * r2 + r2**2)
-
-
-def resistive_load_segments(ls, r1, r2, dr):
-    """Resistive load using truncated cone approximation."""
-    segment_integrals = np.empty_like(ls)
-    is_constant = np.isclose(dr, 0)
-
-    # Constant-radius segments: simple cylinder
-    segment_integrals[is_constant] = ls[is_constant] / (r1[is_constant] ** 2)
-
-    # Varying-radius segments: truncated cone integral
-    segment_integrals[~is_constant] = (
-        ls[~is_constant]
-        / dr[~is_constant]
-        * (1 / r1[~is_constant] - 1 / r2[~is_constant])
-    )
-
-    return np.sum(segment_integrals) / np.pi
-
-
-def average_radius(lengths, r1, r2):
-    """Average radius weighted by segment length."""
-    return np.sum((r1 + r2) / 2 * lengths) / np.sum(lengths)
-
-
-def compute_cone_props(ls, rs, l_start=None, l_end=None):
-    """
-    Given positions ls and radii rs along a path, compute average radius,
-    surface area, volume, resistive load at start and end segments.
-
-    Returns:
-        avg_r: float
-        surface_area: float
-        volume: float
-        res_in: float
-        res_out: float
-    """
-    # Handle default bounds
-    l1 = ls[0] if l_start is None else l_start
-    l2 = ls[-1] if l_end is None else l_end
-
-    # TODO: Add handling spherical compartments? Check if this even needs a special case?
-
-    assert l1 < l2, "Invalid integration bounds"
-
-    # Add boundary points via interpolation
-    l1_l2 = jnp.array([l1, l2])
-    r1_r2 = np.interp(l1_l2, ls, rs)
-    i1_i2 = np.searchsorted(ls, l1_l2)
-    ls = np.insert(ls, i1_i2, l1_l2)
-    rs = np.insert(rs, i1_i2, r1_r2)
-
-    # Select segment within bounds
-    mask = (l1 <= ls) & (ls <= l2)
-    l_seg = ls[mask]
-    r_seg = rs[mask]
-
-    # add midpoint
-    lmid = (l1 + l2) / 2
-    rmid = np.interp(lmid, ls, rs)
-    mid = np.searchsorted(l_seg, lmid)
-    l_seg = np.insert(l_seg, mid, lmid)
-    r_seg = np.insert(r_seg, mid, rmid)
-
-    # Compute segment lengths and properties
-    dl = np.diff(l_seg)
-    r1s = r_seg[:-1]
-    r2s = r_seg[1:]
-    dr = r2s - r1s
-
-    # Compute quantities
-    surface_area = np.sum(surface_area_segments(dl, r1s, r2s, dr))
-    volume = np.sum(volume_segments(dl, r1s, r2s))
-    avg_r = average_radius(dl, r1s, r2s)
-    res_in = resistive_load_segments(dl[:mid], r1s[:mid], r2s[:mid], dr[:mid])
-    res_out = resistive_load_segments(dl[mid:], r1s[mid:], r2s[mid:], dr[mid:])
-
-    return avg_r, surface_area, volume, res_in, res_out
-
-
 def radius_from_xyzr(
     xyzr: np.ndarray,
     min_radius: Optional[float],
